cli bookmark: check only present remote refs
Various commands check if the bookmarks they operate on have a tracking
remote ref. If they do, an error or warning is generated. In all of
these cases, the error or warning only makes sense if the remote
bookmark is actually present.

Until recently, a remote tracking but absent bookmark was not possible
this code. That explains why there is code that implicitly assumes that
remote tracking bookmarks are present.

The affected commands are:
- git push
- abandon
- bookmark rename
- bookmark set
- bookmark create